Java Programming
Create a class Date with the variables to represent the day, month and year, use constructor to initialize the variables (use default date 01/01/1800), mutator and accessor methods, method to return the difference between two Date in days, method to add or subtract n number of days, toString method to return object as a string, equals method to check two Date objects are equal or not. Also validate the date in constructor and mutator methods. [Hint: Use private helper methods]